% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_get_data_as_tibble.R
\name{aoc_get_data_as_tibble}
\alias{aoc_get_data_as_tibble}
\title{Get Advent of Code data as a tibble}
\usage{
aoc_get_data_as_tibble(year, day)
}
\arguments{
\item{year}{Numeric. Year of the challenge.}

\item{day}{Numeric. Day of the challenge.}
}
\value{
Tibble containing data from the wanted \strong{year} and \strong{day}.
}
\description{
Obtains the data from the wanted \strong{year} and \strong{day} and transform the data into a tibble.

Reading the response into a tibble makes the following assumptions using \code{\link[readr:read_delim]{readr::read_delim()}};
\itemize{
\item delim = "\\n"
\item skip_empty_rows = FALSE
\item show_col_types = FALSE
}

Responses will be cached to reduce the number of requests sent to the Advent of Code website.
}
\examples{
\dontrun{
d <- aoc_get_data_as_tibble(2022, 1)
}
}
\seealso{
Other HTTP Request functions: 
\code{\link{aoc_get_response}()}
}
\concept{HTTP Request functions}
