% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_get_response.R
\name{aoc_get_response}
\alias{aoc_get_response}
\title{Get a response from Advent of Code}
\usage{
aoc_get_response(year, day)
}
\arguments{
\item{year}{Numeric. Year of the challenge.}

\item{day}{Numeric. Day of the challenge.}
}
\value{
A \strong{httr2_response} object that will need to be interacted with using \href{https://httr2.r-lib.org/reference/index.html#handle-the-response}{httr2's response handling functions}.
}
\description{
Get the data from the wanted year and day for Advent of Code. Responses will be cached to reduce the number of requests sent to the Advent of Code website.
}
\examples{
\dontrun{
resp <- aoc_get_response(2022, 1)
}
}
\seealso{
Other HTTP Requests: 
\code{\link{aoc_get_data_as_tibble}()}
}
\concept{HTTP Requests}
