% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_set_session_id.R
\name{aoc_set_session_id}
\alias{aoc_set_session_id}
\title{Assist in obtaining and setting the session ID}
\usage{
aoc_set_session_id()
}
\value{
Nothing.
}
\description{
If the session ID is already in possession, use this function to correctly set the environment variable in the \strong{.Renviron} file. If the session ID is not known, answer and follow the prompts in order to navigate and obtain the session ID.
}
\examples{
\dontrun{
aoc_set_session_id()
}
}
