% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc_view_puzzle.R
\name{aoc_view_puzzle}
\alias{aoc_view_puzzle}
\title{Open the puzzle in the browser}
\usage{
aoc_view_puzzle(year, day)
}
\arguments{
\item{year}{Numeric. Year of the challenge.}

\item{day}{Numeric. Day of the challenge.}
}
\value{
Nothing.
}
\description{
Opens a web browser to the page of the wanted \strong{year} and \strong{day}
}
\examples{
\donrun{
aoc_view_puzzle(2022, 1)
}
}
