% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aocr-package.R
\docType{package}
\name{aocr-package}
\alias{aocr}
\alias{aocr-package}
\title{aocr: An R client for accessing Advent of Code’s data}
\description{
This package will only help obtain data, it will not assist in posting solutions via HTTP request. All requests are cached to reduce the number of interactions with Advent of Code’s website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KoderKow/aocr}
  \item \url{https://koderkow.github.io/aocr/}
  \item Report bugs at \url{https://github.com/KoderKow/aocr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Harris \email{koderkow@gmail.com}

}
\keyword{internal}
