% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Get Advent of Code input}
\usage{
input(
  year = curr_year(),
  day = curr_day(),
  path = paste0("input", day, ".txt"),
  verbose = TRUE,
  open_file = FALSE
)
}
\arguments{
\item{year}{Contest year. The default value is current day.}

\item{day}{Day of the puzzle. The default value is current day.}

\item{path}{File path to save input data.
The default vale is "input\{day\}.txt".}

\item{verbose}{Logical flag for printing HTTP response.
The default value is \code{TRUE}.}

\item{open_file}{Logical flag for open to saved file.
The default value is \code{FALSE}.}
}
\value{
The file path where input saved.
}
\description{
By default the \code{input()} downloads current date
\href{https://adventofcode.com/}{Advent of Code} puzzle input.
Also can be used for previous puzzles.
}
\examples{
\dontrun{
input()
input(2021, 1)
purrr::walk(1:25, ~input(2020, .x))
}
}
